kernel: ARM port can build and generate a uImage file now

+++ src/system/include/pedigree/kernel/utilities/lib.h
@@ -145,13 +145,13 @@ EXPORTED_PUBLIC size_t prevCharacter(const char *s, size_t i);
 uint8_t checksum(const uint8_t *pMemory, size_t sMemory);
 
 /// Fletcher 16-bit checksum.
-uint16_t checksum16(const uint8_t *pMemory, size_t sMemory);
+uint16_t checksum16(const uint16_t *pMemory, size_t sMemory);
 
 /// Fletcher 32-bit checksum.
-uint32_t checksum32(const uint8_t *pMemory, size_t sMemory);
+uint32_t checksum32(const uint32_t *pMemory, size_t sMemory);
 
 /// Fletcher 32-bit checksum.
-uint32_t checksum32_naive(const uint8_t *pMemory, size_t sMemory);
+uint32_t checksum32_naive(const uint32_t *pMemory, size_t sMemory);
 
 /// Checksum a page of memory.
 uint32_t checksumPage(uintptr_t address);
